requirements.txt
django==5.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
websockets==12.0
redis==5.0.1
daphne==4.0.0